requirements.txt
scipy
PyWavelets
faiss-cpu
tqdm
PyYAML
imbalanced-learn
scikit-learn
joblib
matplotlib